requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0
daphne==4.2.1  # ASGI server
gunicorn==22.0.0
whitenoise==6.7.0
python-decouple==3.8
